backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
httpx==0.25.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
firebase-admin==6.2.0
psycopg2-binary==2.9.9
pytest==7.4.3
alembic==1.12.1
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
jinja2==3.1.2
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
aiohttp==3.8.6 